requirements.txt
basepage~=1.7.4
requests~=2.27.1
pytest-selenium~=3.0.0
pytest-zebrunner~=1.5.6
